Fix #56 SIP headers are limited to 255 characters
The buffer limits for SIP headers To and From in incoming requests and in-dialog messages were removed.
The buffer length for SIP headers To, From and Refer-To in outgoing requests has doubled.
The buffer length for URL parameters in the WMS App has doubled.
The other conditions remained due to the need for major restructuring.